Change svg tag position to be on its own line.
The reason for this change is because our code always used to depend on svg tag being at the very end as well as it being on a new line, so we could capture it with a conditional. I also believe this is the reason for our maps not rendering.